Make .release-notes support optional
When I added support for adding release notes from the release-notes-bot,
it was intended to be optional. However, because the rotation of the
new-release.md file wasn't wrapped in a check to see if it exists,
then using 0.2.1 without .release-notes/new-release.md would error out.

After this commit, release-notes-bot support is now actually optional.