Make ASSET_NAME hardcoded
Previously, you had to provide an ASSET_NAME env variable to the
announcement script. However, this was always the name of the repo.

With this change, we now use the repo name itself rather than
requiring a name to be given. This will change most things slightly
as the repo name contains the org.

Previously an asset name might be "corral" but would now be "ponylang/corral"

Closes #5