Allow for overriding how trigger release announcement step gets version information
Prior to this commit, the trigger release announcement step
gets the version from GITHUB_REF. For this to work, the step
has to be triggered by a tag being pushed.

This is good for most use-cases, but not all. For example, ponyc
build assets for release on cirrus-ci not github actions. This
means that we want to trigger a release announcement once we've
been notified assets have been uploaded to cloudsmith.

This will not work with trigger release announcement as it can't
get the version from GITHUB_REF as it isn't available there.

This commit adds the ability to provide an optional environment
variable VERSION to the trigger-release-announcement step. If the
variable is set, the version will be taken from that rather than
trying to parse it out of GITHUB_REF.

Closes #22