Replace "step" input parameter with overriding of entrypoint
GitHub actions already provides a mean to run a specific "subcommand"
via changing the entrypoint. Having a "step" input that operated in the
same way but different from what GitHub actions had built in feels problematic.

This change is a first step towards a rather large redoing of the release-bot-action.