Switch supported actions/checkout from v1 to v2
When we were supporting v1, we had to take a default branch to push
changes back to as well as jump through hoops with setting up proper
access urls to push to it.

actions/checkout v2 is a lot more featureful. One of those features
is to take a branch (under the input `ref`) to check out as well as
the token to use to checkout.

Switching to v2 and moving the "default branch" and "token to push/pull
GitHub with" into the checkout step from our scripts is an overall win.

This is a breaking change as I don't want to try to figure out supporting
both v1 and v2.

Closes #32