Update examples to not directly pull the docker image
It doesn't work because the default branch default value isn't picked
up. Thanks "bug" in GitHub actions.

Prior to the next version being released, this usage will still result
in the image being used as the action.yml will be updated on release
to use the image and will then be put back to using the "build it
every time Dockerfile".