Switch Dockerfile to be based on Ubuntu 20.04
PyGitHub needs the python crypto library and there are no wheels for
that available on Alpine. To use the newer version of PyGitHub, we need
to build the crypto library from source and that now means setting up
rust.

The result... it takes much longer to build from an Alpine base.

Closes #49