Handle release already existing when uploading release notes
Previously, the uploading of release notes assumed that the release
didn't exist yet. This made sense for all our ponylang usage. However,
for other projects that create assets as part of the build process
and store them in GitHub releases rather than Cloudsmith, the release
might already exist.

Now, we will check to see if the release exists. If it does, we update
it with our release notes. If it doesn't, we will create the release
like we previously did.